requirements.txt
asgiref==3.6.0
bcrypt==4.0.1
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.1.7
django-cors-headers==3.14.0
djangorestframework==3.14.0
idna==3.4
mysqlclient==2.1.1
PyJWT==2.6.0
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.2
sib-api-v3-sdk==7.5.0
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.15
